move NoopQuerySegmentWalker and add document
